Don’t seek way_idx if position is already correct
Seeking flushes BufWriter<>, so it is better to reduce them.

Note that delete_way() is reading/writing way_idx to clean way_data,
which needs to seek, so adding a flush. We want to remove this deletion
when importing a new file, which is detected by the fact that way.idx is
too small to contain way_id.